Doubly linked list

+++ DataStructures/DArray.h
@@ -7,9 +7,8 @@
 #ifndef DARRAY_H
 #define DARRAY_H
 
-#include <initializer_list>
+ #include <initializer_list>
 #include <iostream>
-#include <cassert>
 
 template <typename T>
 class DArray
@@ -118,7 +117,7 @@ public:
     // Req 001 This function shall return the current size of the array
     int Size()
     {
-        return m_backIndex;
+        return static_cast<int>(m_backIndex);
     }
 
     // Req 001 This function shall return the element at the specified index if the specified index is valid
